# api/openapi/current.json
{
  "components": {
    "schemas": {
...
      "McpLearningPlan": {
        "properties": {
          "evidenceCount": {
            "minimum": 0,
            "type": "integer"
          },
          "proposalVersion": {
            "minimum": 1,
            "type": "integer"
          },
          "proposedSteps": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array"
          },
          "schemaVersion": {
            "enum": [
              1
            ],
            "type": "integer"
          },
          "usage": {
            "additionalProperties": true,
            "type": "object"
          }
        },
        "required": [
          "schemaVersion",
          "proposedSteps",
          "usage",
          "evidenceCount",
          "proposalVersion"
        ],
        "type": "object"
      },
      "McpSearchResult": {
        "properties": {
          "query": {
            "type": "string"
          },
          "schemaVersion": {
            "enum": [
              1
            ],
            "type": "integer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/RetrievalSourceResponse"
            },
            "type": "array"
          }
        },
        "required": [
          "schemaVersion",
          "query",
          "sources"
        ],
        "type": "object"
      },
...
      "RetrievalCitationResponse": {
        "properties": {
          "sourceUrl": {
            "format": "uri",
            "type": "string"
          },
          "timestampSeconds": {
            "nullable": true,
            "type": "number"
          }
        },
        "required": [
          "sourceUrl",
          "timestampSeconds"
        ],
        "type": "object"
      },
      "RetrievalSourceResponse": {
        "properties": {
          "citation": {
            "$ref": "#/components/schemas/RetrievalCitationResponse"
          },
          "content": {
            "type": "string"
          },
          "score": {
            "type": "number"
          },
          "sourceId": {
            "type": "string"
          },
          "sourceKind": {
            "enum": [
              "post",
              "course_step"
            ],
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "visibility": {
            "enum": [
              "private",
              "public"
            ],
            "type": "string"
          }
        },
        "required": [
          "sourceKind",
          "sourceId",
          "visibility",
          "title",
          "content",
          "score",
          "citation"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "enrollment": {
        "in": "cookie",
        "name": "studytube_enrollment",
        "type": "apiKey"
      },
      "internalApiKey": {
        "in": "header",
        "name": "x-internal-api-key",
        "type": "apiKey"
      },
      "mcpService": {
        "bearerFormat": "MCP service assertion",
        "scheme": "bearer",
        "type": "http"
      },
      "session": {
        "in": "cookie",
        "name": "studytube_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Authenticated StudyTube web and learning API",
    "title": "StudyTube API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/internal/mcp/learning/plan": {
      "post": {
        "operationId": "McpController_createLearningPlan",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpLearningPlan"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "mcpService": []
          }
        ],
        "tags": [
          "Mcp"
        ]
      }
    },
    "/internal/mcp/search": {
      "post": {
        "operationId": "McpController_search",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpSearchResult"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "mcpService": []
          }
        ],
        "tags": [
          "Mcp"
        ]
      }
    },
    "/internal/mcp/tool-calls": {
      "post": {
        "operationId": "McpController_recordToolCall",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "accepted": {
                      "enum": [
                        true
                      ],
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "accepted"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "mcpService": []
          }
        ],
        "tags": [
          "Mcp"
        ]
      }
    },
...
  },
  "security": [
    {
      "session": []
    }
  ],
  "servers": [],
...
}